requirements.txt
uvicorn
FastAPI
paho-mqtt >= 2
python-